+++ tsup.base.config.ts
@@ -1,3 +1,4 @@
+import { type Plugin } from 'esbuild';
 import { type Options } from 'tsup';
 import lightningCssPlugin from 'unplugin-lightningcss/esbuild';
 
@@ -9,8 +10,8 @@ const env = process.env.NODE_ENV || 'development';
 const { npm_package_version } = process.env;
 const istanbulPredicate: Predicate = args => defaultPredicate(args) && !/\.worker\.[cm]?[jt]s$/u.test(args.path);
 
-type Plugin = NonNullable<Options['plugins']>[number];
-const disablePlugin = (pluginName: string): Plugin => ({
+type EsbuildPlugin = NonNullable<Options['plugins']>[number];
+const disablePlugin = (pluginName: string): EsbuildPlugin => ({
   name: `disable-plugin-${pluginName}`,
   esbuildOptions: options => {
     const plugin = options.plugins?.find(({ name }) => name === pluginName);
@@ -20,34 +21,47 @@ const disablePlugin = (pluginName: string): Plugin => ({
   }
 });
 
-const cssPlugin = lightningCssPlugin({
-  include: [/\.module\.css$/u],
-  options: {
-    cssModules: {
-      pattern: 'w[hash]_[local]',
-      pure: true,
-      animation: false,
-      grid: false,
-      customIdents: false
-    }
+// Due to a bug in unplugin, we cannot use the Lightning CSS unplugin directly.
+// https://github.com/unjs/unplugin/issues/546
+// To workaround the bug, we are converting it back to esbuild plugin so we can apply a custom `filter`.
+// Otherwise, unplugin will apply `filter: /.*/` and trigger the bug down the road.
+const cssPlugin: Plugin = {
+  name: 'lightningcss',
+  setup: build => {
+    lightningCssPlugin({
+      include: [/\.module\.css$/u],
+      options: {
+        cssModules: {
+          pattern: 'w[hash]_[local]',
+          pure: true,
+          animation: false,
+          grid: false,
+          customIdents: false
+        }
+      }
+    }).setup({
+      ...build,
+      // eslint-disable-next-line require-unicode-regexp
+      onLoad: (_filter, fn) => build.onLoad({ filter: /(\.module_built\.css|\?css_module)$/ }, fn)
+    });
   }
-});
+};
 
 function applyConfig(
   overrideOptions: (
     options: Omit<Options, 'entry' | 'onSuccess'> & {
       define: Record<string, string>;
-      esbuildPlugins: Plugin[];
+      esbuildPlugins: EsbuildPlugin[];
       target: Target[];
     }
   ) => Options & {
     define: Record<string, string>;
-    esbuildPlugins: Plugin[];
+    esbuildPlugins: EsbuildPlugin[];
     target: Target[];
   }
 ): Options & {
   define: Record<string, string>;
-  esbuildPlugins: Plugin[];
+  esbuildPlugins: EsbuildPlugin[];
   target: Target[];
 } {
   const nextOptions = overrideOptions({